webapp/requirements.txt
Click==7.0
Flask==1.1.1
gcloud==0.17.0
googleapis-common-protos==1.6.0
httplib2==0.13.1
itsdangerous==1.1.0
Jinja2==2.10.1
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycryptodome>=3.4.3
Pyrebase==3.0.27
python-dotenv==0.10.3
python-jwt==2.0.1
requests==2.20.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.12.0
Werkzeug==0.15.5
oauthlib==3.1.0
olefile==0.46
pandas==1.0.1
pandocfilters==1.4.2
pickleshare==0.7.5
requests-oauthlib==1.3.0
seaborn==0.10.0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
google==2.0.3
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-auth-oauthlib==0.4.1
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
gunicorn==20.0.4
idna==2.9
pytz==2020.1
urllib3==1.25.9
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==1.5.0
matplotlib==3.1.3
numpy==1.18.1
numpydoc==0.9.2
pyzmq==19.0.0
gcsfs==0.6.2

